move sendMessage to containers as NewMessage
